# swagger.json
{
   "swagger": "2.0",
   "info": {
     "title": "Food-Delivery",
     "description": "<b>RESPONSES</b>::\n 200-> OK\n   401->Not Authorized\n   201->No Record\n 400->Bad Request \n 500->Internal Server Error",
     
       "version": "1.0"
   },
   "produces": ["application/json"],
   "paths": {
...
     "/api/mobile/profile/getProfile": {
      "get": {
        "x-swagger-router-controller": "/api/mobile/profile/getProfile",
        "operationId": "index",
        "tags": ["Get Profile"],
        "description": "Get Profile",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"}

     

     ],

   
    
    
        "responses": {}
      }
    },
    "/api/mobile/profile/updateProfile": {
      "post": {
        "x-swagger-router-controller": "/api/mobile/profile/updateProfile",
        "operationId": "index",
        "tags": ["Update Profile"],
        "description": "Update Profile",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
          { "name": "firstName",  "in": "formData",  "required":true },
          { "name": "lastName",  "in": "formData"},
          { "name": "email",  "in": "formData",  "required":true },
          { "name": "profileImage",  "in": "formData",  "required":true ,"type":"file"},
          { "name": "address",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "dob",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "maritalStatus",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "anniversaryDate",  "in": "formData","type":"string"}


     ],

   
    
    
        "responses": {}
      }
    },
...
   }
}